Fb 1.2.8 snapshot (#15)
* Adding bluegreen flag to the plugin.

* Adding bluegreen flag to the plugin.

* Playing with bluegreen flag in the plugin. Trying to learn what gets used and when.

* Playing with bluegreen flag in the plugin. Trying to learn what gets used and when. Adding back setting labkey.bluegreen using systemProperty.

* Updating flag to testValidationOnly. This should now be similar in style to other gradle flags and more meaningful to other tests. Conditional on dropDatabase to not happen if testValidationOnly is true.

* Updating readme.

* Updating ReadMe to include date.

* Updating with feedback from code review.

+++ src/main/groovy/org/labkey/gradle/plugin/TeamCity.groovy
@@ -229,8 +229,14 @@ class TeamCity extends Tomcat
                         task.dbPropertiesChanged = true
                         task.fn = {
                             properties.mergePropertiesFromFile()
-                            if (extension.dropDatabase)
-                                SqlUtils.dropDatabase(project, properties)
+                            if (extension.dropDatabase) {
+                                if ((Boolean) extension.getTeamCityProperty("testValidationOnly")){
+                                    logger.info("The 'testValidationOnly' flag is true, not going to drop the database.")
+                                }
+                                else {
+                                    SqlUtils.dropDatabase(project, properties)
+                                }
+                            }
                             properties.interpolateCompositeProperties()
                         }
                         task.doLast {
